requirements.txt
Django==1.10.1
django-countries==4.0
django-localflavor==1.3
django-tastypie==0.13.3
gunicorn==19.6.0
python-dateutil==2.5.3
python-mimeparse==1.5.2
six==1.10.0
